small refactoring
